added working POST test for peak-routes
