+++ lab-yana/route/peak-route.js
@@ -8,9 +8,9 @@ const Mountains = require('../model/mountains.js');
 
 const peakRouter = module.exports = new Router();
 
-peakRouter.put('/api/mountains:mountainsID/peak', jsonParser, function(req, res, next) { //update existing peak info in Mountains
-  debug('PUT: /api/mountains:mountainsID/peak');
+peakRouter.post('/api/mountains/:mountainsID/peak', jsonParser, function(req, res, next) { //update existing peak info in Mountains
+  debug('POST: /api/mountains/:mountainsID/peak');
   Mountains.findByIdAndAddPeak(req.params.mountainsID, req.body, { new: true } )
-  .catch(next)
-  .then(peak => res.json(peak));
+  .then(peak => res.json(peak))
+  .catch(next);
 });
